backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
alembic
openai
anthropic
pypdf
python-docx
python-multipart
playwright
requests
